Organize ClawAuth into topic directories
ClawAuth was the only module keeping files at its root, and its ChatGPT
directory grouped by vendor rather than by role: the refreshing credential
actor the daemon runs, the device flow a login walks, and the pinned wire
constants everything reads all shared one folder.

Six directories now, each holding one role. Commands carries the three
workflows plus the bootstrap and lock lease a command needs before it may
act; Presentation carries the terminal seam and the exit-code table; OAuth,
Credentials, Catalog, and Wire split what ChatGPT held.

ChatGPTModelPicker moves out of ChatGPTModelCatalog.swift into its own file.
The picker is a pure function over an already-fetched catalog, with no HTTP,
no metadata, and no parsing; login calls it twice and never touches the
parser beside it. Its tests split the same way.

Everything else is a file move. SwiftPM globs this target, so Package.swift
is unchanged.
