+++ Sources/ClawAuth/Catalog/ModelSelection.swift
