+++ Sources/ClawAuth/Commands/AuthLogoutWorkflow.swift
