+++ Sources/ClawAuth/Commands/AuthMutationCoordinator.swift
