+++ Sources/ClawAuth/Commands/AuthStatusWorkflow.swift
