+++ Sources/ClawAuth/Credentials/ChatGPTCredentialSource.swift
