+++ Sources/ClawAuth/Credentials/ChatGPTTokenMetadata.swift
