+++ Sources/ClawAuth/OAuth/ChatGPTOAuthModels.swift
