+++ Sources/ClawAuth/Presentation/AuthCommandResultMapper.swift
