+++ Sources/ClawAuth/Presentation/AuthTerminal.swift
