+++ Sources/ClawAuth/Wire/ChatGPTProviderMetadata.swift
