+++ Sources/ClawAuth/Wire/ChatGPTWireValues.swift
