+++ Tests/ClawAuthTests/Commands/AuthBootstrapTests.swift
