+++ Tests/ClawAuthTests/Commands/AuthLockingTests.swift
