+++ Tests/ClawAuthTests/Commands/AuthStatusWorkflowTests.swift
