+++ Tests/ClawAuthTests/Credentials/ChatGPTCredentialShutdownTests.swift
