+++ Tests/ClawAuthTests/Credentials/ChatGPTCredentialSourceTests.swift
