+++ Tests/ClawAuthTests/Credentials/ChatGPTTokenMetadataTests.swift
