+++ Tests/ClawAuthTests/Wire/ChatGPTProviderMetadataTests.swift
